3.02

+++ src/main/java/net/redewhite/lvdataapi/modules/ReceptorCreator.java
@@ -1,10 +1,14 @@
 package net.redewhite.lvdataapi.modules;
 
 import net.redewhite.lvdataapi.developers.API;
+import net.redewhite.lvdataapi.developers.events.receptors.ReceptorDeleteEvent;
+import net.redewhite.lvdataapi.developers.events.receptors.ReceptorLoadEvent;
+import net.redewhite.lvdataapi.developers.events.receptors.ReceptorSaveEvent;
+import net.redewhite.lvdataapi.developers.events.receptors.ReceptorUnloadEvent;
 import net.redewhite.lvdataapi.receptors.InactiveVariable;
 import net.redewhite.lvdataapi.receptors.ActiveVariable;
-import net.redewhite.lvdataapi.types.ConnectionType;
 import net.redewhite.lvdataapi.DataAPI;
+import org.bukkit.Bukkit;
 import org.bukkit.plugin.Plugin;
 
 import java.sql.PreparedStatement;
@@ -15,6 +19,7 @@ import java.sql.ResultSet;
 import java.util.List;
 
 import static net.redewhite.lvdataapi.DataAPI.*;
+import static net.redewhite.lvdataapi.developers.API.getVariableValue;
 
 @SuppressWarnings("unused")
 public class ReceptorCreator {
@@ -51,12 +56,8 @@ public class ReceptorCreator {
             }
         }
 
-        String query = table.getDatabase().getConnectionType().getInsertQuery();
-        query = ConnectionType.replace(query, table.getBruteID(), name, getBruteID(), getDate(), table.getDatabase().getBruteID());
-
-        String query2 = table.getDatabase().getConnectionType().getSelectQuery();
-        query2 = ConnectionType.replace(query2, "id", table.getBruteID(), "WHERE bruteid = '" + bruteID + "'", table.getDatabase().getBruteID());
-
+        String query = table.getDatabase().getConnectionType().getInsertQuery(table.getBruteID(), name, getBruteID(), table.getDatabase().getBruteID());
+        String query2 = table.getDatabase().getConnectionType().getSelectQuery("id", table.getBruteID(), "WHERE bruteid = '" + bruteID + "'", table.getDatabase().getBruteID());
         try {
             try (ResultSet result2 = table.getDatabase().createStatement().executeQuery(query2)) {
                 if (!result2.next()) {
@@ -65,7 +66,7 @@ public class ReceptorCreator {
                     }
                 }
 
-                query = ConnectionType.replace(table.getDatabase().getConnectionType().getSelectQuery(), "*", table.getBruteID(), "WHERE bruteid = '" + bruteID + "'", table.getDatabase().getBruteID());
+                query = table.getDatabase().getConnectionType().getSelectQuery("*", table.getBruteID(),  "WHERE bruteid = '" + bruteID + "'", table.getDatabase().getBruteID());
                 try (ResultSet result = table.getDatabase().createStatement().executeQuery(query)) {
                     if (result.next()) {
                         ResultSetMetaData rmtd = result.getMetaData();
@@ -81,6 +82,9 @@ public class ReceptorCreator {
             e.printStackTrace();
             getReceptors().remove(this);
         }
+
+        ReceptorLoadEvent event = new ReceptorLoadEvent(this);
+        Bukkit.getPluginManager().callEvent(event);
     }
 
     public TableCreator getTable() {
@@ -100,6 +104,13 @@ public class ReceptorCreator {
         return variables;
     }
 
+    public boolean hasLoadedBefore() {
+        return getVariableValue("hasLoadedBefore", this).asBoolean();
+    }
+    public long timesLoaded() {
+        return getVariableValue("timesLoaded", this).asLong();
+    }
+
     public void save() {
         StringBuilder query = new StringBuilder();
 
@@ -110,11 +121,20 @@ public class ReceptorCreator {
         }
 
         query.append("last_update = '").append(getDate()).append("'");
-        String fQuery = table.getDatabase().getConnectionType().getUpdateQuery();
-        fQuery = ConnectionType.replace(fQuery, table.getBruteID(), query.toString(), getBruteID(), table.getDatabase().getBruteID());
+        String fQuery = table.getDatabase().getConnectionType().getUpdateQuery(table.getBruteID(), query.toString(), getBruteID(), table.getDatabase().getBruteID());
         table.getDatabase().executeQuery(fQuery);
+
+        ReceptorSaveEvent event = new ReceptorSaveEvent(this);
+        Bukkit.getPluginManager().callEvent(event);
     }
     public void unload() {
+        ReceptorUnloadEvent event = new ReceptorUnloadEvent(this);
+        Bukkit.getPluginManager().callEvent(event);
+
+        if (event.isCancelled()) {
+            return;
+        }
+
         save();
         getReceptors().remove(this);
 
@@ -123,6 +143,13 @@ public class ReceptorCreator {
     }
 
     public void delete() {
+        ReceptorDeleteEvent event = new ReceptorDeleteEvent(this);
+        Bukkit.getPluginManager().callEvent(event);
+
+        if (event.isCancelled()) {
+            return;
+        }
+
         try (PreparedStatement pst = table.getDatabase().getConnection().prepareStatement("DELETE FROM '" + table.getBruteID() +  "' WHERE bruteid = '" + bruteID + "';")) {
             pst.execute();
             unload();
